Another try to cope with warnings under gcc 8.2.

+++ utils/funcexp/func_substring_index.cpp
@@ -74,7 +74,7 @@ std::string Func_substring_index::getStrVal(rowgroup::Row& row,
     if ( count > (int64_t) end )
         return str;
 
-    if (( count < 0 ) && ((count * -1) > end))
+    if (( count < 0 ) && ((count * -1) > (int64_t) end))
         return str;
 
     string value = str;
